+++ CHANGELOG.markdown
@@ -8,6 +8,11 @@
 
 * fix no retorno do valor inputado na hora da branch
 
+Bug:
+
+* symlinks: calling setup_vars to ensure path's.
+* setup: cleaning up releases based on old time mask.
+
 ## 0.2.0
 
 New: